modularization and refactor of gql
TODO restructure project so gql is sibling of api
