+++ api/graph_search/interfaces.go
@@ -1,0 +1,15 @@
+package graphsearch
+
+import "blockbuster/api/data"
+
+type MovieGraphInterface interface {
+	BFS(start string, stars, movieIDs, directors map[string]bool, depth int)
+	GetDirectedActors(director string) []string
+	GetDirectedMovies(director string) []data.Movie
+	GetStarredIn(star string) []data.Movie
+	GetStarredWith(star string) []string
+	GetMoviesByID(ids map[string]bool) []data.Movie
+	TotalStars() int
+	TotalMovies() int
+	TotalDirectors() int
+}
